requirements.txt
matplotlib
python-docx
numpy
